Adding battery monitor sketch
This is a simple sketch I'm using to monitor the
voltage of a test battery over time.  Essentially this
sketch just reads off ADC0 at regular intervals and
converts that into a voltage reading.  That value
is then printed onto the serial port, which can
be logged by a host computer to get a graph of voltage
over time.  Use the collect_logs.sh script to add
timestamps and save them to the host computer's hdd.

Signed-off-by: Rave Strapon <ravestrapon@gmail.com>