Use internal ADC reference for battery monitor
Using the internal 2.56V reference voltage instead
of Vcc for measuring the battery voltages.  This makes it
much more accurate, but requires the use of external
voltage dividers to bring the signals into range, so the
SW has to compensate.  This patch also includes precisely
computed scalers for each of the probes on my battery
monitor -- obviously those numbers have to change with
different resistor values.

Signed-off-by: Rave Strapon <ravestrapon@gmail.com>