Adding in the first draft of the actual src
This is heavily copied from another project of mine
but most importantly includes a build.sh that can compile
and flash using my ATmega16 board.  This is tricky because
FastLED needs Arduino, and I'm not using Arduino to update
nor is there an ATmega16 Arduino.  So this requires:

MightyCore installed in your Arduino install.  This adds ATmega16
support for Arduino.  It's pretty easy to install.

A slightly hacked version of FastLED -- I just used the existing
fastpin.h file but modified the line that referenced the ATmega16U2
and made it ATmega16 instead.  (Also remove the HOODBOOT condition)
I was surprised to see it work, but sure enough, it seems to work
like that.

Then since there is no usb/serial on the board we can't use the
Arduino bootloader to update the code, so the build script
just compiles a .hex file that includes the Arduino bootloader
and my code.  Then I use an ISP programmer to flash the whole thing
on in one go.

This is hacky as heck, but it does indeed work for the time being
at least!

Signed-off-by: Rave Strapon <ravestrapon@gmail.com>