Add connection test to build script
This adds a simple connection test to the build script so it
can give more informative errors.  Without this, when you don't
have the programmer or mcu plugged in, the error messages are
about fuses which can be confusing.  This way it fails earlier
and it's clear you just need to check your connections and the
error has nothing to do with setting fuses.

Signed-off-by: Rave Strapon <ravestrapon@gmail.com>