Purge outdated references to pole/ball in animations
The original project I wrote these animations for was a vertical
art installation with a large ball on top of a pole.  As such these
animations had to control these two kinds of lights, but is now
overkill since the new design has only a "pole" (so-to-speak).

This CL just cleans up the animation code and gets rid of these
confusing references.

Signed-off-by: Rave Strapon <ravestrapon@gmail.com>