Remove erroneous 1s pause
For some reason there was a 1s pause after each pass
through the animations.  This caused a weird glitch
each time it restarted at the top of the list again.

This patch just removes this delay.  It looks much smoother
without it and I think it was just a forgotten leftover
piece of a test or something.

Signed-off-by: Rave Strapon <ravestrapon@gmail.com>