Add support for buttons & Fuel gauge
This is rudamentary support for a "fuel gauge" on ADC4, which
is meant to be wired to a 1/2 voltage divider connected to the
battery voltage.

The two buttons are wired to external interrupts 0 & 1 so they
can be interrupt-driven buttons, which is also new.  The code is
somewhat janky for this and can't interrupt the animations yet
(which kinda defeats the whole point) but I'm working for fix that
currently.

Right now when you press the brightness button the brightnesses update
on the fly, but the fuel gauge won't display until the current
animation is done.  I'll have to refactor the animation code to
achieve that.

Signed-off-by: Rave Strapon <ravestrapon@gmail.com>