Refactor for interruptable animations
Previously there was no good was to interrupt an animation due
to the way they were implimented.  This new OOP method should
fascilitate a more regular frame rate and allow you to interrupt
them which is required for a fuel gauge to work right.

Only a couple animations are refactored now.  More to follow

Signed-off-by: Rave Strapon <ravestrapon@gmail.com>