Adding a frame count to animations
This allows the animation to be totally selfcontained -- now
the nextFrame() function returns a bool (indicating if it's
generated as many frames as expect already) and the main loop
can just move on when it says false.  This means we can just push
a whole new animation into place and it'll restart right back
where it expected.  This will allow interrupting animations (like
for the battery indicator) much much more smoothly

Signed-off-by: Rave Strapon <ravestrapon@gmail.com>