Add an interrupting Fuel Guage
The fuel guage now "works" although the animation is lame right now.
When you press the button a Fuel Guage animation is pushed into place
over whatever the current animation is and then run out before it
returns to that same animation seamlessly.  No more waiting for an
animation to complete before the battery status is revealed!  Next, I
need to make it look better, check if power is plugged in, and possibly
stay on as long as you hold down the battery button.

Signed-off-by: Rave Strapon <ravestrapon@gmail.com>