Keep the fuel guage on as long as the btn is down
This is a slick user interface upgrade.  As long as the user
holds down the button the battery level will keep being displayed.
If you don't hold it down though, it'll just play for a little bit
and then stop, do you don't *have* to hold it down if you don't
want to.  Also this adds a pulsing animation to the fuel guage that
greatly upgrades it's appearance

Signed-off-by: Rave Strapon <ravestrapon@gmail.com>