Add in USB power detection to the fuel gauge.
This CL adds in the logic to the fuel gauge to make sure that the USB
power cable isn't plugged in when it's doing it's readings.  If it is,
then it plays the "currently charging" animation instead of the "fuel
gauge" one, to let the user know that it's plugged in.  The animation
isn't the greatest right now, but the plumbing is all in place.

Still to do:
    Make it play an animation automatically when you plug/unplug
    the usb power to let you know it's working.

    Better animation for charging

    Hook up the status LEDs from the LiPo charger to GPIOs to get
    better/accurate charging status information.

Signed-off-by: Rave Strapon <ravestrapon@gmail.com>