Make the system more robust to errors
This is accomplished by enabling a watchdog timer.  The WDT reboots
the microcontroller if it doesn't get reset for 60ms.  This shouldn't
happen unless something is awry though -- because the WDT gets reset
right before attempting to generate a new frame.  There should definitely
be a new frame generated at least once every 60ms, so if that isn't
happening it makes sense to reboot the whole microcontroller.

To try and make this rebooting more "seamless" I also added some code
to store the brightness settings to EEPROM and resore them on boot.
This way, if you had the brightness down very low and the system gets
rebooted by the WDT, you won't suddenly have the brightness jump up
to whatever the default value is.  You might notice the pattern glitch/
jump to the next one suddenly, but the brightness value should stay the
same, which is easily the most visible way it can screw up.

Signed-off-by: Rave Strapon <ravestrapon@gmail.com>